+++ internal/dsn/dsn_test.go
@@ -36,11 +36,34 @@ func TestParseConnectionString(t *testing.T) {
 			"123",
 		},
 		{
+			"grpc://ydb-ru.yandex.net:2135/ru/home/gvit/mydb?token=123",
+			false,
+			"ydb-ru.yandex.net:2135",
+			"/ru/home/gvit/mydb",
+			"123",
+		},
+		{
 			"grpcs://ydb.serverless.yandexcloud.net:2135/?" +
 				"database=/ru-central1/b1g8skpblkos03malf3s/etn02qso4v3isjb00te1&token=123",
 			true,
 			"ydb.serverless.yandexcloud.net:2135",
 			"/ru-central1/b1g8skpblkos03malf3s/etn02qso4v3isjb00te1",
+			"123",
+		},
+		{
+			"grpcs://ydb.serverless.yandexcloud.net:2135" +
+				"/ru-central1/b1g8skpblkos03malf3s/etn02qso4v3isjb00te1?token=123",
+			true,
+			"ydb.serverless.yandexcloud.net:2135",
+			"/ru-central1/b1g8skpblkos03malf3s/etn02qso4v3isjb00te1",
+			"123",
+		},
+		{
+			"grpcs://ydb.serverless.yandexcloud.net:2135" +
+				"/ru-central1/b1g8skpblkos03malf3s/etn02qso4v3isjb00te1?database=/ru/home/gvit/mydb&token=123",
+			true,
+			"ydb.serverless.yandexcloud.net:2135",
+			"/ru/home/gvit/mydb",
 			"123",
 		},
 		{
@@ -52,7 +75,22 @@ func TestParseConnectionString(t *testing.T) {
 			"123",
 		},
 		{
+			"grpcs://lb.etn03r9df42nb631unbv.ydb.mdb.yandexcloud.net:2135" +
+				"/ru-central1/b1g8skpblkos03malf3s/etn03r9df42nb631unbv?token=123",
+			true,
+			"lb.etn03r9df42nb631unbv.ydb.mdb.yandexcloud.net:2135",
+			"/ru-central1/b1g8skpblkos03malf3s/etn03r9df42nb631unbv",
+			"123",
+		},
+		{
 			"abcd://ydb-ru.yandex.net:2135/?database=/ru/home/gvit/mydb",
+			true,
+			"ydb-ru.yandex.net:2135",
+			"/ru/home/gvit/mydb",
+			"",
+		},
+		{
+			"abcd://ydb-ru.yandex.net:2135/ru/home/gvit/mydb",
 			true,
 			"ydb-ru.yandex.net:2135",
 			"/ru/home/gvit/mydb",